+++ builder/qemu/step_http_ip_discover.go
@@ -2,8 +2,11 @@ package qemu
 
 import (
 	"context"
+	"fmt"
+	"net"
 
 	"github.com/hashicorp/packer/helper/multistep"
+	"github.com/hashicorp/packer/packer"
 )
 
 // Step to discover the http ip
@@ -12,7 +15,52 @@ import (
 type stepHTTPIPDiscover struct{}
 
 func (s *stepHTTPIPDiscover) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
-	state.Put("http_ip", "10.0.2.2")
+	config := state.Get("config").(*Config)
+	ui := state.Get("ui").(packer.Ui)
+
+	hostIP := ""
+
+	if config.NetBridge == "" {
+		hostIP = "10.0.2.2"
+	} else {
+		bridgeInterface, err := net.InterfaceByName(config.NetBridge)
+		if err != nil {
+			err := fmt.Errorf("Error getting the bridge %s interface: %s", config.NetBridge, err)
+			ui.Error(err.Error())
+			return multistep.ActionHalt
+		}
+		addrs, err := bridgeInterface.Addrs()
+		if err != nil {
+			err := fmt.Errorf("Error getting the bridge %s interface addresses: %s", config.NetBridge, err)
+			ui.Error(err.Error())
+			return multistep.ActionHalt
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip == nil {
+				continue
+			}
+			ip = ip.To4()
+			if ip == nil {
+				continue
+			}
+			hostIP = ip.String()
+			break
+		}
+		if hostIP == "" {
+			err := fmt.Errorf("Error getting an IPv4 address from the bridge %s: cannot find any IPv4 address", config.NetBridge)
+			ui.Error(err.Error())
+			return multistep.ActionHalt
+		}
+	}
+
+	state.Put("http_ip", hostIP)
 
 	return multistep.ActionContinue
 }