+++ post-processor/vsphere-template/post-processor.go
@@ -19,6 +19,7 @@ import (
 	"github.com/hashicorp/packer/helper/config"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
+	"github.com/hashicorp/packer/post-processor/artifice"
 	"github.com/hashicorp/packer/template/interpolate"
 	"github.com/vmware/govmomi"
 )
@@ -27,6 +28,7 @@ var builtins = map[string]string{
 	vspherepost.BuilderId:  "vmware",
 	vmwcommon.BuilderIdESX: "vmware",
 	vsphere.BuilderId:      "vsphere",
+	artifice.BuilderId:     "artifice",
 }
 
 type Config struct {