+++ provisioner/powershell/provisioner.go
@@ -98,13 +98,13 @@ type Provisioner struct {
 
 func (p *Provisioner) defaultExecuteCommand() string {
 	baseCmd := `& { if (Test-Path variable:global:ProgressPreference)` +
-		`{set-variable -name variable:global:ProgressPreference -value 'SilentlyContinue'};`
+		`{set-variable -name variable:global:ProgressPreference -value 'SilentlyContinue'};` +
+		`{set-variable -name variable:global:ErrorActionPreference -value 'Continue'};`
 
 	if p.config.DebugMode != 0 {
 		baseCmd += fmt.Sprintf(`Set-PsDebug -Trace %d;`, p.config.DebugMode)
 	}
-
-	baseCmd += `. {{.Vars}}; &'{{.Path}}'; exit $LastExitCode }`
+	baseCmd += `. {{.Vars}};try { & '{{.Path}}' } catch { Write-Error $Error[0]; exit 1 }; if ($LastExitCode) { exit $LastExitCode }}`
 
 	if p.config.ExecutionPolicy == ExecutionPolicyNone {
 		return baseCmd
@@ -371,10 +371,11 @@ func (p *Provisioner) createRemoteCleanUpCommand(remoteFiles []string) (string, 
 		return "", fmt.Errorf("clean up script %q failed to upload: %s", remotePath, err)
 	}
 
-	data := map[string]string{
-		"Path": remotePath,
-		"Vars": p.config.RemoteEnvVarPath,
-	}
+	data := p.generatedData
+	data["Path"] = remotePath
+	data["Vars"] = p.config.RemoteEnvVarPath
+	p.config.ctx.Data = data
+
 	p.config.ctx.Data = data
 	return interpolate.Render(p.config.ExecuteCommand, &p.config.ctx)
 }